showAllPosts - V

+++ react-project/src/components/Post.jsx
@@ -20,29 +20,40 @@ const Post = ({ post, setPosts, posts }) => {
     }));
   }
 
-  function deleteTodoClicked() {
-    const requestOptions = {
-      method: 'DELETE',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(copyPost)
-    };
+  function deletePostClicked() {
+    if (user.id != post.userId)
+      alert("you can not delete this post :(");
+    else {
+      const requestOptions = {
+        method: 'DELETE',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(copyPost)
+      };
 
-    fetch(`http://localhost:3000/posts/${copyPost.id}`, requestOptions)
-      .then(() => {
-        let i, updatePosts;
-        posts.map((t, index) => {
-          if (t.id === copyPost.id) {
-            i = index;
-          }
-          return t;
+      fetch(`http://localhost:3000/posts/${copyPost.id}`, requestOptions)
+        .then(() => {
+          let i, updatePosts;
+          posts.map((t, index) => {
+            if (t.id === copyPost.id) {
+              i = index;
+            }
+            return t;
+          });
+          updatePosts = [...posts]
+          updatePosts.splice(i, 1);
+          setPosts(updatePosts);
+        })
+        .catch((error) => {
+          console.error('There was an error!', error);
         });
-        updatePosts = [...posts]
-        updatePosts.splice(i, 1);
-        setPosts(updatePosts);
-      })
-      .catch((error) => {
-        console.error('There was an error!', error);
-      });
+    }
+  }
+
+  function editPostClicked() {
+    if (user.id != post.userId)
+      alert("you can not edit this post :(");
+    else
+      setEditState(true);
   }
 
   function handleSubmit() {
@@ -81,25 +92,50 @@ const Post = ({ post, setPosts, posts }) => {
             setEditState(false);
           navigate(`/home/users/${user.id}/posts/${post.id}`);
         }}>{post.title}</h2>
-        <button onClick={deleteTodoClicked}>🚽</button>
+        <button onClick={deletePostClicked}>🚽</button>
         <button onClick={() => {
-            navigate(`/home/users/${user.id}/posts/${post.id}/comments`),
+          navigate(`/home/users/${user.id}/posts/${post.id}/comments`),
             setViewComment((prev) => !prev)
-          }}>🍾</button>
+        }}>🍾</button>
       </div>
-      {viewComments && <Outlet  /> }
+      {viewComments && <Outlet />}
       {formPost && <div className='postDetails'>
         <h2>{copyPost.id}</h2><br />
-        <input type="text" name="title" className='input' value={copyPost.title} onChange={handleChange} disabled={!editState} /><br />
-        <input type="text" name="body" className='input' value={copyPost.body} onChange={handleChange} disabled={!editState} /><br />
-        {!editState && <>
-          <button onClick={() =>
-            setEditState(true)}>🪁</button>
-        </>}
-        {editState && <>
+        {/* <input type="text" name="title" className='input' value={copyPost.title} onChange={handleChange} disabled={!editState} /><br />
+        <input type="text" name="body" className='input' value={copyPost.body} onChange={handleChange} disabled={!editState} /><br /> */}
+
+        <label>
+          Post title:
+          <input
+            name="title"
+            disabled={!editState}
+            value={copyPost.title}
+            onChange={handleChange} />
+        </label>
+        <br />
+        <label>
+          <textarea
+            name="body"
+            disabled={!editState}
+            value={copyPost.body}
+            onChange={handleChange}
+            rows={6}
+            cols={50}
+          />
+        </label>
+
+
+        {/* {!editState && <>
+          <button onClick={editPostClicked}>🪁</button>
+        </>} */}
+        {/* {editState && <>
           <button onClick={handleSubmit}>✅</button>
           <button onClick={() => { setFormPost(false) }}>❎</button>
-        </>}
+        </>} */}
+        <hr />
+        {!editState &&<button onClick={editPostClicked}>Edit</button>}
+        {editState && <><button onClick={handleSubmit}>Save Post</button>
+          <button>Reset edits</button></>}
       </div>}
     </>
   );
